Call events from ExecuteVerb handler instead of direct functions
The reason is so the target object can take advantage of component local events. There was a conflict between stairs and the world map entrances, for example. If a StairsComponent is available but a WorldMapEntranceComponent isn't, then the StairsSystem shouldn't do anything.

The problem is ExecuteVerbEventArgs operates on the verb user, not the target, so the target's components are ignored.