Make submodules use HTTPS instead of Git URLs
I was originally intending for these to be Git so that you could make changes to submodules inside the OpenNefia repo, but using branches and `--remote` means the HEAD is detatched, so it becomes impractical. But Git URLs also mean people can't even clone the repo unless they set up an SSH key in their GitHub account (if they even have one).