Weave broadcast/component events together and sort by default
Gets rid of the old priority system. The issue is that it becomes hard to manually specify before/after handlers every time if there are system event handlers that must go first/last. Also it's hard to reference event identifiers in external libs with the SubId system unless the IDs are always made available as static vars.

Numeric priority has been used in many other games (as well as ON/Love) and works reasonably well.