Always serialize EntityUids from outside the map in full map serialization
Problem: Entity sets a personal faction relation against the player and an ally. Those entity UIDs are saved in a dictionary of <EntityUid, Relation>. However, because the player's party is not in the map at the time of (full mode) saving, the serializer complains that they're not part of the map and the UID becomes `null`. Hence a duplicate key exception.